Fix for heatlink var
